requirements.txt
Flask==3.1.0
Pillow==11.1.0
numpy==2.0.2
ultralytics==8.3.68
ultralytics-thop==2.0.14
opencv-python
textwrap